+++ CHANGELOG.md
@@ -14,6 +14,7 @@ All notable changes to this project will be documented in this file.
 - Confirma uses invariant culture when working with strings & numbers.
 - TestDiscovery class is now static.
 - Renamed protected '_autoload' & '_output' fields to 'Autoload' & 'Output'.
+- Repeat field in TestCase is now of type RepeatAttribute and not ushort.
 
 ## [0.6.1-beta 2024-07-08]